+++ ChangeLog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Version [0.6.0](https://github.com/cedx/html.net/compare/v0.5.0...v0.6.0)
+- Added the `New-TrackElement` cmdlet.
+- Added the `track` alias.
 
 ## Version [0.5.0](https://github.com/cedx/html.net/compare/v0.4.0...v0.5.0)
 - Added the `New-ColGroupElement`, `New-DataElement`, `New-StyleElement` and `New-VideoElement` cmdlets.